T-7: Activity Form (replaces AddEntryModal)
- Add components/activity-form.tsx — Dialog on desktop, Drawer on mobile (useIsMobile hook via matchMedia)
- Tag multi-select via Popover with checkboxes + inline tag creation; loads tags from getAllActivityTags() on open
- Pre-populates tags from editItem.tags in edit mode (uses ActivityWithRelations instead of Activity)
- Inline POC and venue type creation retained from old modal
- Calls createActivity/updateActivity with tagIds array
- Recurrence logic preserved (toggle + frequency + occurrence count preview)
- Delete components/add-entry-modal.tsx; update CalendarDaySidebar and DayViewClient to use ActivityForm
